tests/requirements.txt
pytest>=7.0.0
boto3>=1.26.0
hypothesis>=6.0.0
PyYAML>=6.0
requests>=2.28.0
dnspython>=2.3.0
beautifulsoup4>=4.11.0